+++ src/VaultSharp/V1/AuthMethods/Token/ITokenAuthMethod.cs
@@ -10,6 +10,18 @@ namespace VaultSharp.V1.AuthMethods.Token
     public interface ITokenAuthMethod
     {
         /// <summary>
+        /// Creates a new token. 
+        /// Certain options are only available when called by a root token. 
+        /// If you are creating an orphaned token, a root token is not required to create an orphan token 
+        /// (otherwise set with the no_parent option). 
+        /// If used with a role name, the token will be created against the specified role name; 
+        /// this may override options set during this call.
+        /// </summary>
+        /// <param name="createTokenRequest">The token creation request</param>
+        /// <returns>Auth info</returns>
+        Task<Secret<object>> CreateTokenAsync(CreateTokenRequest createTokenRequest);
+
+        /// <summary>
         /// Gets the calling client token information. i.e. the token used by the client as part of this call.
         /// </summary>
         /// <returns>